#117 file names and commit history fix
