+++ src/controllers/helpers/password_module.js
